Remove occurrences of str.format from importlib
It uses `str.format` for cases where format string literals work fine and require no runtime support.
One use was left in `WindowsRegistryFinder` since it is a public member. Since Windows support is not
important to us right now, no need to worry about it being in the way of implementing importlib.

Based on Facebook D9442800