New C++ representation of Objects
Rather than passing around raw `Object`*s, turn `Object` into a
thin wrapper around a `uword` that we pass by value. This fixes two significant
issues with the current system:
- Our liberal use of `reinterpret_cast<word>(this)` for the immediate types is
  a subtle bug waiting to happen. C++ compilers can assume that `this`
  is never null, and optimizations made under that assumption can interact
  poorly with code that pays too much attention to the actual value of
  `this`. Instead, represent an `Object` as an opaque 64-bit value that we cast
  to a pointer when appropriate, getting rid of all casts from `this`.
- Code like `Handle<Object> foo(...); foo->setBar(newThing());` is not GC-safe
  if `newThing()` can allocate. `foo.operator->()` is called first, creating a
  temporary `Object*`. Then, if `newThing()` triggers a GC, the pointee of that
  `Object*` can move, causing a crash when `Object::func()` is finally
  called. In the new system, `foo.operator->()` still returns an `Object*`, but
  it points to the `Object` inside the `Handle` that can't move, not the actual
  heap object.

I've added an `operator->` to all of the Object classes; this is meant to be
temporary to reduce the footprint of this diff, and I will remove them in a
series of followup diffs.

Based on Facebook D9927060