Handle dunder call methods bound to descriptors
It is possible to encounter a non-data descriptor object when calling
through the dunder call method of a callable object.  This requires
evaluating the dunder get method and retrying the call method with
its return value.  Because the return value itself can be a callable
object this process must be repeated until either a function is found
or some kind of error occurs.

Based on Facebook D9358585