Change frame layout
This commit changes the frame layout in two major ways:

1. Arguments will no longer be copied down the stack from the
   caller's frame into the callee.
2. The dynamic portion of the stack is no longer reserved as part
   of frame allocation. The interpreter is expected to use the thread
   stack to push and pop values.

As a result of (1), we'll need to rewrite LOAD_FAST into a pair of
new bytecodes (LOAD_ARG and LOAD_LOCAL?). In CPython, the localsplus
member of the frame is an array that contains slots for the arguments,
local variables, cell variables, and free variables, in that order.
We're effectively splitting localsplus into two (arguments and
the rest); LOAD_ARG will handle the first half, and LOAD_LOCAL will
handle the remainder.

Canonical stack frames now look like:

  Top of stack / higher memory addresses
  [Locals N]
  ...
  [Locals 0]
  <Other fixed size frame metadata>
  [Globals]
  [Builtins]
  [Code object]
  [Saved frame pointer]  <- Frame pointer
  [Arg N]
  ...
  [Arg 0]
  [Function]

Based on Facebook D6129027