Fix tests that assume LayoutId creation order
Some tests in `runtime-tests.cpp` assume that the first user-defined class will get the LayoutId right after the last builtin.
This breaks if we define any new builtins in python frozen modules that are loaded by the runtime.

Change these tests to ask for the layout id of the type or the instance in order to avoid this assumption.

Based on Facebook D9373844