Pyro: Extend python::Vector to accept custom allocators
This extends python::Vector to accept custom allocators,
addressing a pending TODO.

This also moves out the malloc and free calls from the Vector class
into a new SimpleAllocator class which is compliant with the
C++ Allocator concept: http://en.cppreference.com/w/cpp/concept/Allocator

Note: The standard library encourages the use of the helper
allocator_traits to provide a common interface to all allocators. This
though, adds an extra dependency to the memory library. We could remove
it and directly call the allocator member.

Based on Facebook D7314506