[pyro] Add List Insert
This adds an insert function for list objects.

How it works:
This uses listAdd() to guarantee a space to insert the new element.
Then, we correct the insert index in case it is out of bounds i.e. [0, Current Total].
After that, we move all the object pointers larger than the insert position to their pos + 1.
Finally, we insert our object in the correct position.

Based on Facebook D7363757