Refactor old-style string formatting tests (second try)
Use runtime.runFromCString instead of testing the helper method. It's a bit simpler; it will work even when we move part of the logic to Python

**Tests on D9412550 were failing in -release mode, which I didn't
notice. This is because Str::charAt does only a DCHECK_INDEX on the
index and the current implementation didn't check, so the behavior
with invalid format was different. Adding a simple fix to the current
implementation.**

Based on Facebook D9438165