[Pyro] C-API: Convert boolobject-test to only use the C-API
This modified boolobject-test.cpp to use the CAPI only.

This also adds a test fixture which is in charge of starting and stopping the runtime through the pylifecycle functions. Given this, we can now add tests that use the CAPI that completely lack any reference to the runtime headers. This will make the tests more similar to what external code does.

Based on Facebook D9347247