[Pyro] Remove use of ApiTypeHandle
This is the first change towards fully treating PyTypeObject as an implementation detail in the Runtime.

This change fully removes ApiTypeHandle and makes PyTypeObjects go through the reference pointer just like any ApiHandle.

The next step is to mark PyType_Ready as `UNHANDLED` and move all that code to PyType_FromSpec. This will help to fully remove our dependency on PyTypeObject.

Based on Facebook D9420420