[Pyro] Builtin: Implement a simple len()
This adds the `len` builtin.

How does it work?
It currently just checks if an object is a list. If so, it casts it to a List and obtains the allocated() value. There's a TODO to build a better way to identify all the objects that support len (i.e. strings, tuples, dicts, etc)

Why?
This builtin comes in handy when testing list.pop()

Based on Facebook D7379708