Declare no-argument functions as f(void) but define them as f()
In C++ the declarators f() and f(void) both mean a no-argument
function.  However, in C, f() is a function of unknown arguments and
f(void) is a function that takes no arguments.

This change uses f(void) in places visible to both a C and C++
compiler and f() in places visible only to a C++ compiler.

Based on Facebook D10211549