Ensure nan != nan
Summary: Since floating point NaNs have the property of (NaN !=x; for any value of x), this needs to be tested. This was brought up in the context of D10286141. This does not check that this does not break hash table backed datatypes like dict, set, frozenset. See: https://repl.it/@bsimmers/py-nan.

Based on Facebook D10559130