[Pyro] Add List Pop Builtin
This adds the list.pop builtin.

First makes a check to verify that we are popping from a list.
If no argument is specified, default to list->allocated().
This also adds in some index handling to check out of bounds and negative indexing as well.
Then, call listPop(...)

Based on Facebook D7379481