Support subclassing built-in types that are subclasses of built-in types
In order to define `ImportError` in Python, we need to support subclassing
of built-in types that are themselves subclasses of a built-in type. In
this particular case, `ImportError` subclasses `Exception`, and `Exception`
subclasses `BaseException`.

We know that if the candidate (`Exception` in this case) mro contains all
the other built-in classes in the computed mro then there can be no lay-out
conflict.

Based on Facebook D9309209