Wildcard the .gitignore for CLion build directories.
The latest release of CLion supports multiple toolchains.  Too keep
the build artifact directories for each toolchain separate, the
name of the toolchain is added to the directory name.  Ignoring
anything that starts with "cmake-build-" should be sufficient to
capture all permutations of this naming scheme.

Based on Facebook D6461136