Support for objects with count >= 255.
This change adds an optional overflow word that immediate preceeds
objects whose count is greater than 254.  The overflow word contains a
tagged small integer with the count and the header count field is set
to a flag value of 255 signaling to look at the preceeding word for
the count.

This change also offsets all object pointers positively by one word.
An object's address is now the first word of the body instead of the
first (or second) word of the header.

Based on Facebook D6234264