native exception API
Add API methods to Thread which will support throwing Python
exceptions from native code. The implementation of those functions
is currently a placeholder, until we add exception support to the
interpreter.

The contract for native code is similar to that required by CPython.
In CPython, throwing a python exception from native code involves
calling a method to post the exception, and returning NULL from the
native entry point. This API also requires calling a method (on Thread)
to post the exception, and then an object of type Error must be
returned to the native entry point.

Additionally, there is now a common trampoline method template for
all native code. This eliminates the alignment requirement on native
function pointers, and gives us a place outside of the CALL_FUNCTION
fast path to check if native exceptions should be thrown.

Based on Facebook D6613733