Implement LOAD_GLOBAL and STORE_GLOBAL instructions.
Global variables are implemented.  This is done through a level of
indirection so we can cache the locations of global variables in
code to avoid the hash-lookups at runtime.

This change also adds some additional infrastructure around modules
including creating the sys and __main__ modules.  Hashing is also
done implicitly in some dictionary interfaces.  Realistically, the
dictionaries manipulated by the runtime will not invoke user-defined
hashing functions.  That will have to be handled at a much higher
level.  When ready, we will likely expose the underlying dictionary
lookup method.

Based on Facebook D6388631