Vector and StackVector implementations
Our own vector implementation, based on llvm's SmallVector

std::vector<> is almost good enough, but its missing a few things that can
really help us:

* The ability to use stack storage, thus avoiding mallocs, when we know
we're only going to have a small handful of elements. This is already done.
* The ability to tune our memory allocator. std::vector obviously supports
allocators, but this will be easier to work with and more flexible.

Not supported yet: non-POD types. Our common case is likely to be vectors of
pointers, so this may be good enough for a long time.

Based on Facebook D6551382