Add API to create and execute builtin functions
Function objects grew a new `builtin` attribute that, if set, should
contain a pointer to a C function with the following signature:

  Object* (*)(Thread*, Frame*, Object** argv, word argc, Object* kwNames)

This signature is a little onerous on the part of the implementor but
should be general enough to handle the different ways in which a function
may be called. One thing to note is that `argv` points at the *last*
argument since it is stack allocated. Most of the builtins
(https://docs.python.org/3/library/functions.html) take a small number of
arguments, so I don't think this will be too painful.

Add an implementation of `print()` that will allow us to execute the
"Hello World" program.

Based on Facebook D6238654