Test running a program starting from the module function.
Python programs are started from top-level functions that are
compiled into a module function.  This test runs the hello world
function called fro the module function.

This test identified a few places where the global environment was not
being propogated into its location in callee frames.  Either because a
global was not being assigned to a function at creation time or
because the specially crafted open frame for module procedures did not
set it up in the first place.

Based on Facebook D6438240