Create a table for mapping class IDs to class objects.
This is a first step toward having user defined classes.  All class
objects are now stored in a single table and a class ID doubles as
an index into the class table where the class object for that ID
is located.

The bottom 32 entries in the class table are reserved for the classes
of immediate types like SmallInteger.  This will allow a class
object lookup for an immediate type (which has no header, and
therefore no storage location for a class ID) to use the bottom 5
bits of the object as a table index.

A future change will add new entries to the end of the table to
allocate a class ID and store its associated class object.

Finally, this change adds some infrastructure for a future mmediate
string object type.

Based on Facebook D6493772