Add function creation, implicit globals, and print.
This change adds the remaning function to make the "hello, world"
function to run in the unit test harness.

In addition to adding the MAKE_FUNCTION, LOAD_NAME, and STORE_NAME
opcodes, the print function is added to the builtins module and a
slot for implicit globals have been added to stack frames.  Some unused
slots from frame objects were removed.

Based on Facebook D6421488