Add rudimentary instance size calculation
This diff adds support for computing the size (the number of attributes
) of class instances. It does so by scanning the constructors of each
member of a class's MRO looking for sequences of bytecodes of the form

  LOAD_FAST   0  // Load self
  STORE_ATTR  N  // Set the attribute named co_names[N]

There are some limitations of the current implementation:

1. It performs a linear scan of the constructors bytecode. As a result,
   it will incorrectly compute the instance size of the following class
   as 2, when it should be 1.

       class Foo:
         def __init__(self):
           if True:
             self.foo = None
           else:
             self.bar = None
2. It assumes that each class invokes the constructor of its superclasses.

Both of these limitations will only result in an overestimate of instance
size.

I also created a dictionary for each class that is manually created by
the runtime. A separate diff will deal with refactoring the class
creation boilerplate.

Based on Facebook D6612562