Eliminate direct usage of SmallString/LargeString. Clients must use the String interface.
We don't want every usage of the String API to have to distinguish between
large and small strings. Doing so spreads needless complexity throughout the code base,
amplifies the risk that particular code is only correct for one class of strings, and
increases the burden placed on adding future storage methods to String, e.g.  Ropes.

Based on Facebook D6760925