MRO computation for user-defined classes.
MRO computation for user-defined classes.

This diff adds computation of MRO for user-defined classes using
C3 linearization. The implementation is O(n^2), as is CPython's.

In the case where no valid MRO can be found, the runtime terminates.
This will be replaced with an exception API real soon.

Based on Facebook D6594677