Add BoundMethod and support for calling non-function objects
In order to get richards working we'll need to support calling both
`BoundMethod` objects (created as a result of calling a method), as well
as class objects (to create instances of user defined classes). This
change adds a `BoundMethod` object and support for calling it from the
interpreter.

Based on Facebook D6788049