Add instance attribute map
Instead of computing only the instance size during class creation, we
also compute a "map" of attribute name to its offset within instances
of the class.

This map will be consulted during attribute lookup with
the tentative plan of falling back on an overflow dictionary when the
map lookup fails.

Initially, we will search the map using a linear scan. This is simple,
and for the majority of classes, it will likely be faster than a more
sophisticated search/indexing technique. Most python classes have a
small number of attributes so once the map is loaded it will likely
reside in one to two cache lines. Additionally, attribute names will
be interned, so we can use pointer comparison for a fast, initial
scan.

Based on Facebook D6837181