Fix stack effect for STORE_FAST
Well this is embarrassing. STORE_FAST consumes a value from the stack,
and since our stack grows down, we should be incrementing the stack
pointer rather than decrementing it.

In order to avoid this in the future we should abstract away stack
manipulation (either behind a set of macros or functions hanging off
of frame).

Based on Facebook D6917885