Basic module support to unblock extension module work.
This change adds basic support for importing modules,
loading attributes from modules, and executing code in
non-__main__ modules. This is the minimum (non-hacky) set of
functionality required for testing C API function calls via
extension modules.

Actually searching for modules on disk is not even attempted here,
but this should provide the foundation for such a change in the
future.

Based on Facebook D6913021