Zero out frames starting from new value stack limit
We track the current stack pointer and the thread's limit pointer
separately. Prior to this change we were using the limit pointer
to zero out frames. Because the limit pointer may lie below the
stack pointer, this resulted in portions of the frame being left
uninitialized.

Based on Facebook D6920629