Make type objects callable.
Instances are allocated in Python by calling the type object.  At
the byte code level, this looks like an ordinary CALL_FUNCTION,
which is typically dispatched to a __call__ method.  This change
takes an interim approach of explicitly checking for a type object
and dispatching to a routine that instantiates an instance and calls
an __init__ method if one is present.

Based on Facebook D6906287