[pyro] Restructure interpreter for IR exctraction
These changes are intended to support experimentation in interpreter
implementation.  In general, this core interpreter function will be
re-implemented to more effectively support interactions with GC and
JIT compilation.

In this diff we restructure the interpreter so each byte code is
implemented by a specific function.  Those functions are then called
through a table instead of a switch statement (although either would
be fine).

The goal was to isolate the semantics of a byte code in a method which
could then be converted directly into an IR suitable for manipulation
by a JIT. A separate tool is write which mechanically translates the
functions in this file into such code. The output of that function is
shown here:

https://phabricator.intern.facebook.com/P58983088

Elements of the execution state, specifically the thread pointer, the
frame pointer, the stack pointer and the current PC are collected into
a context object. The translation assumes this object is the first
parameter and has knowledge of the layout of the structure.

A simple JIT could translate a byte code stream into a sequence of
calls to these functions.  Subsequent JIT refinements could inline
these calls from the extracted high-level IR and perform operations to
eliminate stack pointer operations optimize constant loads, and local
variable accesses.

Based on Facebook D6845497