Add helper to print the current python stack trace to stderr
Add a helper method to `Frame` that dumps the stack trace to stderr.
Additionally, dump the stack trace for all checked aborts.

In order to make this work, we need to sync the virtual pc to the frame
prior to leaving the interpreter.

Sample output from the added test:

```
Traceback (most recent call last)
  File '/tmp/python-tests.Xolwj7t/foo.py', line 11, in <module>
  File '/tmp/python-tests.Xolwj7t/foo.py', line 9, in test
  File '/tmp/python-tests.Xolwj7t/foo.py', line 6, in b
  File '/tmp/python-tests.Xolwj7t/foo.py', line 3, in a
Aborted (core dumped)
```

Based on Facebook D7049880