Add helper script for tracing script execution
This adds a script to help us understand what is required to execute a
piece of code.

When run like so:

  python trace_script.py /path/to/script.py report.html

It first executes the script and records which functions are called.
It then statically analyzes the script to determine what modules are
imported and in what order. Using this information, it writes a report
with the following information:

- The list of modules that were imported.
- The functions that were called, grouped by module and type (interpreted/native).
- A call trace.
- An import trace. Each import occurs exactly once, the first time it was imported.

A sample report produced by running the script on our unittest milestone is available
at https://home.fburl.com/~mpage/unittest_milestone.html.

Based on Facebook D7156944