Add all public C-API funtions to cpython-func.h
This is needed to fork getargs.cpp, D12953145, as the code uses many functions that are not currently in declared in `cpython-func.h`. I don't strictly need to add all of them, but I had to add a few random functions to this file every now and then, so it seems to be better to do this just once and completely.

Add declarations for (almost) all functions defined in `ext/**/*.cpp` and order them alphabetically. This was generated by a python script, P60322697. Types used by the APIs were (manually) moved from `*.cpp` files to `cpython-types.h`.

Based on Facebook D13035585