Remove redefined 'typedef struct' from CPython sources
per discussion in chat, this causes problems with `clang` with pre-C11 standard because of -Wtypedef-redefinition

I'm not convinced this is the best solution - it actually hides a problem where our typdef is incorrect. We current redefine Py_UCS4 as int, while it's uint32_t in CPython sources. Keeping typedefs would reveal this issue (and we can still replace the underlying 'struct' that it's typedefed to). But I don't have a strong preference..

Based on Facebook D13094401