Rework builtin subclass check fast path
Rather than allocating a bit for every class that we want to enable
fast subclass checks for, store the LayoutId of the builtin base type in every
RawType (details of what this means are in objects.h). Now, to check if a type
is a subtype of int, for example, we can just check if its builtin base is
int. The only complication is BaseException, which has subtypes that we also
want to check for. This is handled by putting BaseException and all of its
builtin subtypes in a contiguous range in LayoutId, and doing a range check in
Runtime::isInstanceOfBaseException(). This completely replaces the subclass
flags, so they have been removed and all uses migrated to the new builtin base
system.

Finally, macro-generate the Runtime::isInstanceOfFoo() functions that we've
been writing by hand to reduce duplication, and make sure we have one for all
of the types listed here: https://fburl.com/uoiizz2x.

Based on Facebook D12886031