Rename ObjectArray to Tuple
This has been on our TODO list for a while. The one instance of
ObjectArray that I left as-is is the one in ObjectFormat in objects.h, since
that's describing the internal format of an object rather than naming a
type. If anyone disagrees I can change that one too.

Based on Facebook D12912662