Fork getargs.cpp
This is forking the functionality in `getargs.c` - this is the largest amount of code we want to reuse and some of this might need to be reimplemented in the future anyway. Compiling this source directly (what we do now) might be exposing some unsafe code.

- Copy code from getargs.c to getargs.cpp
- Remove private APIs / functions that are not needed
- clang-format and minimal fixes to make clang-tidy happy
- Keep `PyArg_ValidateKeywordArguments` as unimplemented (it was using another private API), it'll have to be rewritten

With this, we won't need the patch on CPython repo to replace `&PyTuple_GET_ITEM` in this file anymore and since this eliminates one last usage of that pattern, we will finally be able to redefine the macro.

Based on Facebook D12953145