Make managed references in C-API handles distinct
This change introduces a simpler mechanism for tracking borrowed
references.  Instead of distinguishing managed objects that are
borrowed, this change distinguishes managed objects whether or
not they are borrowed.

All managed references have a large count value that is reflected
in the reference count returned by the handle.  This contrasts
with the borrowed status that was kept hidden. Borrowed
references are now just a special case of returning a handle to a
managed object without incrementing its reference count.  All of
the information relating to the borrowed status of a handle has
now been removed.

To support this change, the interfaces for returning a new
reference and a borrowed reference are named by the conceptual
type of reference being returned.  Also, a few tests were updated
to reflect the change in reference count values being returned
for managed objects.

Based on Facebook D12932879