Implement PyOS_snprintf and PyOS_vsnprintf
These wrap the system snprintf and vsnprintf with added behavior.
First, pointer arguments are asserted to be non-NULL and the size
is checked to be non-zero.  Second, when the size is greater than
INT_MAX-1 the system exits early and terminate the buffer with a
'\0'.  The latter case cannot be tested.

Based on Facebook D13148612