Separate out argument processing into leaf subroutines
Previously, argument processing was performed in functions which
performed an interpreter subroutine call.  This meant that the
interpreter subroutine call had to accommodate all combinations of
code objects and argument processing.

With this change, the argument processing for each of the call forms
is separated into its own subroutine than can be called by subroutine
calls specialized to a specific kind of code object and argument
processing.

To demonstrate this effect, the code to handle generators has been
separated from the general interpreter entry points into specialized
generator entry points which unconditionally bottom-out into creating
a new generator.

A further effect of this change is that for interpreter to interpreter
subroutine calls, the argument processing can be be done inside the
interpreter instead of in a recursive invocation of the interpreter.
This is needed for exception handling and will be a part of a
forthcoming change.

Based on Facebook D13090060