Run extension tests with CPython
Our extension tests are meant to test the API contract and shouldn't be implementation-dependent. This enables compiling extension tests with cpython - to test that we're testing the right thing.

This adds running cpython extensions tests to non-asan legocastle jobs as well

This depends on a few open diffs but otherwise should be ready.

Based on Facebook D10148030