Make the clang-tidy lint work again
This makes two changes to enable running clang-tidy on legocastle
and on devservers.

First, the path to the clang-tidy binary has been switch to a
version in third-party2 that is matched to the same version of
the platform that we do our builds with.

Second, to work around a short-term issue with header files in
our build, some problematic C++11 constructs are now
conditionally compiled-out to avoid issues with the older version
of GCC that is provided with CentOS 7.

Based on Facebook D13267660