Implement length/size API functions for abstract protocols
Implement Py{Mapping,Object,Sequence}_{Length,Size} using up-calls to the interpreter.

These can all share a helper because the functions check only for the presence of a `__len__(self)` attribute, and do not fail if the argument does not actually provide the intended protocol.

This is not exactly the same functionality as CPython, in which PySequence_{Length,Size} fail on `dict` but not on subclasses. However, this does mirror the functionality of PyPy.

Based on Facebook D13142445