Only clear AttributeError for getattr/hasattr
`hasattr` is implemented by running `getattr` and checking for an AttributeError. Allow other errors (e.g. TypeError) to pass through the builtin function. Still clear every error for PyObject_HasAttr.

Similarly, `getattr` should only clear the error if a default is provided *AND* the error is an AttributeError.

Based on Facebook D13471635