Allow Python type and doc annotations to supplement existing C++ definitions
Motivation: The native builtin definitions of Python functions are
very barebones. This is not great for debugging or tooling purposes. It would
be great if there was a source of information on how to use them from managed
code. This change allows us to annotate existing native functions in managed
code when bootstrapping a class.

Additionally, the current ad-hoc parameter parsing for native code methods is
scary and needs to be blown away. This diff is a stepping stone toward doing
that; with the annotations in Python, we should be able to automatically parse
out parameters.

Based on Facebook D13089816