Implement the _PyUnicodeWriter API
Summary: The _PyUnicodeWriter API is needed for us to properly implement `PyUnicode_FromFormat`. It's also needed by a handful of other extension modules. However, since we don't rely on PyObjects and switching string encodings, we can effectively implement it using only a buffer we copy into.

Based on Facebook D13439708