Delete testing::createUniqueObject()
It was originally created to give an object of unknown type and value
with a refcount of 1. That's no longer true, since most ApiHandles have the
managed bit set on creation. It was also being misused in a few places: some
tests assumed it wasn't a certain type, some tests assumed it was hashable,
etc.

I replaced most uses of it with `PyLong_FromLong(0)`, because 0 seems like a
reasonable meaningless value and it's cheap to create. The reference counting
tests all operate on relative values now, which should work fine even when 0 is
interned, since we still have to reference count its
`ApiHandle`. `ModuleExtensionApiTest.GetStateallocatesAndAllowsMutation` looks
like it actually wants a guaranteed-unique object, so I used an empty tuple
there.

Based on Facebook D13612870