Delete Runtime::isIteratorExhausted()
This was a hack to get iterators working before we had proper
exception support. Now that we do, it can be removed, with StopIteration being
used as intended in its place.

I also cleaned up a few abortOnPendingException() calls that were in the way of
this cleanup; removing the rest of them is coming next.

Based on Facebook D13616379