Move compilation cache back to Runtime::compile()
Other than what I fixed in D13627426, there are a variety of other
ways to invoke the compiler: `PyRun_SimpleString()`, `exec()` from within
Python, manually calling `Runtime::compile()`, etc. Until we have the compiler
properly integrated into Pyro, put the compilation cache back in the critical
path to cover all of these cases.

This puts us roughly in the state we were in before D13130106, without fully
reverting it. I kept `testing::runFromCStr()` and removed
`Runtime::runFromCStr()`. This eliminates any confusion between the two, and
requires the handful of uses of `Runtime::runFromCStr()` to be more explicit
with what they're doing. We can reevaluate the long-term plan for this cache
when we change cpython versions or start integrating the compiler.

Based on Facebook D13629081