Fix 'str.__getitem__' with negative index
There was a bug in the case that the index is an integer less than zero, so I've added tests for integer indexes.
I also refactored the function to allow for Str and Int subclasses later, and left a breadcrumb for using `__index__` to get the index.

Based on Facebook D13702674