Add argument checking to object methods
Since we can annotate native code with manage code using the boostrapping
classes, this diff automates some of the positionalizing of arguments for the
native code positional trampoline.

It is motivated by the abundance of oft-incorrect ad-hoc argument-checking code
in the builtins classes. With luck, most of that will soon vanish.

Additionally:

* Fix `dict` builtins to use new argument checking trampoline
* Fix `str.__len__` to use new argument checking trampoline

To make a builtin automatically checked, use the builtinTrampoline instead of
the nativeTrampoline and don't bother checking the number of arguments in the
function body.

Based on Facebook D13678275