Create function for invoking methods with interpreter
Rather than making the caller responsible for looking up and then calling the method, add a function to the thread that does both.

Don't set an exception if the lookup fails, so the caller can set its own messages. Preserve exceptions raised by the method call.

Converts a few API functions to show proof-of-concept.

Based on Facebook D13802481