Refactor attributeAt for function, instance
Previously, type attribute lookup did not consider the slots on
the type itself (there were none). This change both adds slots and adds a
lookup step.

Additionally, this modifies functionGetAttr for the same reason: the lookup
process is the same as for instances, except for a dict check. Since the dict
check is common to both function and type, collapse.

The overflowAttributes of a Layout can now be in three states:

1) NoneType; this is a sealed class and there are no overflow attributes
2) SmallInt; this is a class whose "overflow attributes" should be looked up in
   a dict that is stored in a slot on the object
3) Tuple; this is a class with normal overflow attributes

Last, `type.__name__` should be stored in a slot, not in a dict.

Based on Facebook D13795267