Allow slice indices other than SmallInt
In CPython, any object is allow to be a slice index as long as it has a `__index__` method.

Move `RawSlice::unpack` to slice-builtins so we can invoke `__index__` when required and raise ValueError when step is zero.

We should also silently compress LargeInt indexes to min and max values as in CPython.

Based on Facebook D13849634