Implement PyUnicode_Append and PyUnicode_AppendAndDel
Needed for the io module. Follow CPython's behavior.

Acts the same as `PyUnicode_Concat`, and stores the result of the concatenation in the `PyObject** p_left` parameter. It steals the object stores in the `p_left` pointer, borrows the `right` object, and returns a new reference.

Based on Facebook D13772972