Make RawTupleIterator take iterator length
This modifies the interface of RawTupleIterator to also take the length of the tuple that it will iterate. This is in preparation of Struct Sequences which is a subclass of tuple. This type will have a RawTupleIterator with a different length from the real tuple length.

The other alternative is to create a custom iterator for Struct Sequences but that seems overkill when a small modification to RawTupleIterator should be enough.

Based on Facebook D13908245