Drop _thread module for now
Remove `_thread` module for now. Since we don't have any actual support for
multithreading yet, it is better not make it look like we do.

For example `library/_bootstrap.py` has code paths for multithreaded and
singlethreaded pythons. It switches between them based on the presence of
`_thread`.  So not exposing `_thread` will simplify the bringup of importlib.

Based on Facebook D13924992