Fix break, continue, and return implementations to properly handle blockstack
Our implementations of these control-flow operations weren't correctly
inspecting the blockstack. Bring them all up to snuff by adding a helper to pop
one block off the blockstack and using it in all three. As part of this, remove
RETURN_VALUE's special status in the dispatch loop by changing the convention
around opcode handlers that return `bool` (see the updated comment in
`interpreter.h`).

I'm intentionally not implementing full support for finally in this diff, since
this should be enough to get importlib going. That will come in the (hopefully
near) future.

Based on Facebook D13856729