Implement PyObject_Bytes
If the argument is null, then return a bytes representation of it.
If the argument is already bytes, incref and return it.
If the argument has a `__bytes__` method, call it, check that it returns an instance of bytes, and return that.
Otherwise, use bytesFromIterable to convert it from a list, tuple, or generic iterable.

Based on Facebook D13900972