Hide the use of tp_new in posixmodule
posixmodule is one of the few remaining places that actually sets a slot through the object itself rather than letting the type system handle it.

However, for both of these cases, there are workarounds to correctly handle it. For now, this just hides this incompatibility while we wait for a patch to be upstreamed.

Note, this will be a local only change - meaning that the CPython repo that we are tracking does not have these changes. The reason being 1) importlib does not require these to work and 2) this is still not the end solution for these two types. The full solution will be upstreamed to that repo and sync'd here in the future.

Based on Facebook D14022118