Implement PyOS_FSPath
Implements PyOS_FSPath

This also:
* Modifies the scripts that generates the cpython files to also make a pass through posixmodule.c to remove the implementation of PyOS_FSPath
* Deletes PyOS_AfterFork as it's already present in signalmodule.cpp

Please note that posixmodule is one of the few extension modules in CPython that also declares a public C-API. Therefore, adding posixmodule.c to the cpython file generator should not be a common pattern to follow. In general, all extensions should be self contained.

Based on Facebook D14021830