Fix invalid coding pattern in test
This test was creating a 2nd runtime in a callback. Creating a string in a different runtime and
and passing it to another is wrong and only works by accident. Another side effect of this is that after destroying the 2nd runtime the current thread pointer is zero.

The test does not fail at the moment, but I'd like to change it to use a test
helper in a following commit that accesses the current thread pointer.

Based on Facebook D14025346