Reuse function objects from python in builtin-modules
We used to replace the methods in types with `bootstrap=True` with the
previously defined builtin methods and copy over the attributes from the python
side. This failed to take into account other references to the functions. So something like this would not work:

```
class int(bootstrap=True):
  def __int__(self) -> int:
    pass

  real = property(__int__)
```

This changes the algorithm to keep the python function object around and just
copy over the builtin entry points from the native code.

Based on Facebook D14073885